+++ HipWhipGame/Assets/Systems/Scripts/RollbackSupportSystem/Inputs/InputFrame.cs
@@ -6,8 +6,8 @@ namespace RollbackSupport
     public struct InputFrame
     {
         public int frame;
-        public sbyte horiz;
-        public sbyte vert;
+        public float horiz;
+        public float vert;
         public bool light, heavy, grab, block;
     }
